refactor: merge admin nav into sidebar, remove top bar
AdminSidebar is now in AdminLayout (persistent across all admin pages).
AdminNav removed — top bar was redundant with sidebar navigation.
Logout + email moved to sidebar footer.
