+++ src/app/admin/layout.tsx
@@ -2,7 +2,7 @@ import type { ReactNode } from "react";
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 import { checkAdminSession } from "@/lib/auth/session";
-import { AdminNav } from "@/components/admin/AdminNav";
+import { AdminSidebar } from "@/components/admin/AdminSidebar";
 
 export const dynamic = "force-dynamic";
 
@@ -17,9 +17,9 @@ export default async function AdminLayout({ children }: { children: ReactNode })
   const session = await checkAdminSession();
   if (!session) redirect("/admin/login");
   return (
-    <div className="min-h-screen">
-      <AdminNav email={session.email} />
-      {children}
+    <div className="grid min-h-screen" style={{ gridTemplateColumns: "200px 1fr" }}>
+      <AdminSidebar email={session.email} />
+      <main>{children}</main>
     </div>
   );
 }
